Reduce from int32 to uint16 for capacityfactor id maps
Those map simply point to ids of time series, of which we certainly always will have (far) less than 64,000 and hence enough for uint16.

+++ src/capacityfactors/id_map.py
@@ -9,7 +9,7 @@ import rasterio
 from rasterio.transform import from_origin
 import xarray as xr
 
-DTYPE = np.int32
+DTYPE = np.uint16
 NO_DATA_VALUE = -1
 INDEX_EPSILON = 10e-3
